handle eraPoints, rm Subscan provider
