--- src/contexts/Plugins/Subscan/types.ts
@@ -1,8 +1,0 @@
-// Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-import type { AnyJson } from '@polkadot-cloud/react/types';
-
-export interface SubscanContextInterface {
-  fetchEraPoints: (v: string, e: number) => Promise<AnyJson>;
-}
